cleanup: some more MAJOR MINOR unsigned
Finish remaing missing cases where MAJOR/MINOR was casted to (int)
while we can now use it directly as 'unsigned'.

+++ lib/device/dev-io.c
@@ -310,7 +310,8 @@ int dev_open_flags(struct device *dev, int flags, int direct, int quiet)
 
 	if (dm_list_empty(&dev->aliases)) {
 		/* shouldn't happen */
-		log_print_unless_silent("Cannot open device %d:%d with no valid paths.", (int)MAJOR(dev->dev), (int)MINOR(dev->dev));
+		log_print_unless_silent("Cannot open device %u:%u with no valid paths.",
+					MAJOR(dev->dev), MINOR(dev->dev));
 		return 0;
 	}
 	name = dev_name(dev);
